+++ test/UsenetTests/Nntp/Parsers/ListGroupsReponseParserTest.cs
@@ -15,22 +15,22 @@ namespace UsenetTests.Nntp.Parsers
             new object[]
             {
                 211, "1234 3000234 3002322 misc.test list follows", new string[0],
-                new XSerializable<NntpGroup>(new NntpGroup("misc.test", 1234, 3000234, 3002322, NntpPostingStatus.Unknown, string.Empty, new List<int>(0)))
+                new XSerializable<NntpGroup>(new NntpGroup("misc.test", 1234, 3000234, 3002322, NntpPostingStatus.Unknown, string.Empty, new List<long>(0)))
             },
             new object[]
             {
                 211, "1234 3000234 3000236 misc.test list follows", new [] {"3000234", "3000235", "3000236"},
-                new XSerializable<NntpGroup>(new NntpGroup("misc.test", 1234, 3000234, 3000236, NntpPostingStatus.Unknown, string.Empty, new[] {3000234, 3000235, 3000236}))
+                new XSerializable<NntpGroup>(new NntpGroup("misc.test", 1234, 3000234, 3000236, NntpPostingStatus.Unknown, string.Empty, new[] {3000234L, 3000235L, 3000236L}))
             },
             new object[]
             {
                 411, "example.is.sob.bradner.or.barber is unknown", new string[0],
-                new XSerializable<NntpGroup>(new NntpGroup("", 0, 0, 0, NntpPostingStatus.Unknown, string.Empty, new List<int>(0)))
+                new XSerializable<NntpGroup>(new NntpGroup("", 0, 0, 0, NntpPostingStatus.Unknown, string.Empty, new List<long>(0)))
             },
             new object[]
             {
                 412, "no newsgroup selected", new string[0],
-                new XSerializable<NntpGroup>(new NntpGroup("", 0, 0, 0, NntpPostingStatus.Unknown, string.Empty, new List<int>(0)))
+                new XSerializable<NntpGroup>(new NntpGroup("", 0, 0, 0, NntpPostingStatus.Unknown, string.Empty, new List<long>(0)))
             }
         };
